refactor: move statistics utils to libcommon
#TODO: move tests
