+++ services/worker/tests/job_runners/split/test_descriptive_statistics.py
@@ -17,13 +17,7 @@ from huggingface_hub.hf_api import HfApi
 from libcommon.dtos import Priority
 from libcommon.resources import CacheMongoResource, QueueMongoResource
 from libcommon.simple_cache import upsert_response
-from libcommon.storage import StrPath
-
-from worker.config import AppConfig
-from worker.job_runners.config.parquet import ConfigParquetJobRunner
-from worker.job_runners.config.parquet_and_info import ConfigParquetAndInfoJobRunner
-from worker.job_runners.config.parquet_metadata import ConfigParquetMetadataJobRunner
-from worker.job_runners.split.descriptive_statistics import (
+from libcommon.statistics import (
     DECIMALS,
     MAX_NUM_STRING_LABELS,
     MAX_PROPORTION_STRING_LABELS,
@@ -37,10 +31,16 @@ from worker.job_runners.split.descriptive_statistics import (
     ImageColumn,
     IntColumn,
     ListColumn,
-    SplitDescriptiveStatisticsJobRunner,
     StringColumn,
     generate_bins,
 )
+from libcommon.storage import StrPath
+
+from worker.config import AppConfig
+from worker.job_runners.config.parquet import ConfigParquetJobRunner
+from worker.job_runners.config.parquet_and_info import ConfigParquetAndInfoJobRunner
+from worker.job_runners.config.parquet_metadata import ConfigParquetMetadataJobRunner
+from worker.job_runners.split.descriptive_statistics import SplitDescriptiveStatisticsJobRunner
 from worker.resources import LibrariesResource
 
 from ...constants import CI_HUB_ENDPOINT, CI_USER_TOKEN
